new update Hall and Session

+++ Back/CinemaTicketsRealase/src/main/java/com/dao/HallDao.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.query.Param;
 
 public interface HallDao extends JpaRepository <Hall,Long> {
     @Query(value = "select h from Hall as h where h.number = :number")
-    Hall getNumberHall(@Param("number")Integer number);
+    Hall getNumberHall(@Param("number")Long number);
 }
